Extract validation logic for Route into RouteValidator
