+++ internal/k8s/reconciler/route_test.go
@@ -16,8 +16,6 @@ import (
 
 	clientMocks "github.com/hashicorp/consul-api-gateway/internal/k8s/gatewayclient/mocks"
 	"github.com/hashicorp/consul-api-gateway/internal/k8s/reconciler/state"
-	"github.com/hashicorp/consul-api-gateway/internal/k8s/service"
-	"github.com/hashicorp/consul-api-gateway/internal/k8s/service/mocks"
 )
 
 func TestRouteID(t *testing.T) {
@@ -149,167 +147,6 @@ func TestRouteMatchesHostname(t *testing.T) {
 	require.True(t, newK8sRoute(&gwv1alpha2.TCPRoute{}, K8sRouteConfig{
 		Logger: hclog.NewNullLogger(),
 	}).MatchesHostname(&hostname))
-}
-
-func TestRouteValidate(t *testing.T) {
-	t.Parallel()
-
-	require.NoError(t, newK8sRoute(&core.Pod{}, K8sRouteConfig{
-		Logger: hclog.NewNullLogger(),
-	}).Validate(context.Background()))
-
-	require.True(t, newK8sRoute(&gwv1alpha2.HTTPRoute{}, K8sRouteConfig{
-		Logger: hclog.NewNullLogger(),
-	}).IsValid())
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-
-	reference := gwv1alpha2.BackendObjectReference{
-		Name: "expected",
-	}
-	resolved := &service.ResolvedReference{
-		Type:      service.ConsulServiceReference,
-		Reference: &service.BackendReference{},
-	}
-
-	route := newK8sRoute(&gwv1alpha2.HTTPRoute{
-		Spec: gwv1alpha2.HTTPRouteSpec{
-			Rules: []gwv1alpha2.HTTPRouteRule{{
-				BackendRefs: []gwv1alpha2.HTTPBackendRef{{
-					BackendRef: gwv1alpha2.BackendRef{
-						BackendObjectReference: reference,
-					},
-				}},
-			}},
-		},
-	}, K8sRouteConfig{
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-	})
-
-	resolver.EXPECT().Resolve(gomock.Any(), route.GetNamespace(), reference).Return(resolved, nil)
-	require.NoError(t, route.Validate(context.Background()))
-	require.True(t, route.IsValid())
-
-	expected := errors.New("expected")
-	resolver.EXPECT().Resolve(gomock.Any(), route.GetNamespace(), reference).Return(nil, expected)
-	require.Equal(t, expected, route.Validate(context.Background()))
-
-	resolver.EXPECT().Resolve(gomock.Any(), route.GetNamespace(), reference).Return(nil, service.NewK8sResolutionError("error"))
-	require.NoError(t, route.Validate(context.Background()))
-	require.False(t, route.IsValid())
-}
-
-func TestRouteValidateDontAllowCrossNamespace(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-	client := clientMocks.NewMockClient(ctrl)
-
-	//set up backend ref with a different namespace
-	namespace := gwv1alpha2.Namespace("test")
-	route := newK8sRoute(&gwv1alpha2.HTTPRoute{
-		Spec: gwv1alpha2.HTTPRouteSpec{
-			Rules: []gwv1alpha2.HTTPRouteRule{{
-				BackendRefs: []gwv1alpha2.HTTPBackendRef{{
-					BackendRef: gwv1alpha2.BackendRef{
-						BackendObjectReference: gwv1alpha2.BackendObjectReference{
-							Name:      "expected",
-							Namespace: &namespace,
-						},
-					},
-				}},
-			}},
-		},
-	}, K8sRouteConfig{
-		Client:   client,
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-	})
-
-	client.EXPECT().
-		GetReferenceGrantsInNamespace(gomock.Any(), gomock.Any()).
-		Return([]gwv1alpha2.ReferenceGrant{
-			{
-				Spec: gwv1alpha2.ReferenceGrantSpec{
-					From: []gwv1alpha2.ReferenceGrantFrom{},
-					To:   []gwv1alpha2.ReferenceGrantTo{},
-				},
-			},
-		}, nil)
-
-	// FUTURE Assert appropriate status set on route and !route.IsValid() once ReferenceGrant requirement is enforced
-	_ = route.Validate(context.Background())
-}
-
-// TestRouteValidateAllowCrossNamespaceWithReferenceGrant verifies that a cross-namespace
-// route + backend combination is allowed if an applicable ReferenceGrant is found.
-func TestRouteValidateAllowCrossNamespaceWithReferenceGrant(t *testing.T) {
-	t.Parallel()
-
-	ctrl := gomock.NewController(t)
-	defer ctrl.Finish()
-	resolver := mocks.NewMockBackendResolver(ctrl)
-	client := clientMocks.NewMockClient(ctrl)
-
-	//set up backend ref with a different namespace
-	backendGroup := gwv1alpha2.Group("")
-	backendKind := gwv1alpha2.Kind("Service")
-	backendNamespace := gwv1alpha2.Namespace("namespace2")
-	backendName := gwv1alpha2.ObjectName("backend2")
-	route := newK8sRoute(&gwv1alpha2.HTTPRoute{
-		ObjectMeta: meta.ObjectMeta{Namespace: "namespace1"},
-		TypeMeta:   meta.TypeMeta{APIVersion: "gateway.networking.k8s.io/v1alpha2", Kind: "HTTPRoute"},
-		Spec: gwv1alpha2.HTTPRouteSpec{
-			Rules: []gwv1alpha2.HTTPRouteRule{{
-				BackendRefs: []gwv1alpha2.HTTPBackendRef{{
-					BackendRef: gwv1alpha2.BackendRef{
-						BackendObjectReference: gwv1alpha2.BackendObjectReference{
-							Group:     &backendGroup,
-							Kind:      &backendKind,
-							Name:      backendName,
-							Namespace: &backendNamespace,
-						},
-					},
-				}},
-			}},
-		},
-	}, K8sRouteConfig{
-		Client:   client,
-		Logger:   hclog.NewNullLogger(),
-		Resolver: resolver,
-	})
-
-	refGrant := gwv1alpha2.ReferenceGrant{
-		TypeMeta:   meta.TypeMeta{},
-		ObjectMeta: meta.ObjectMeta{Namespace: "namespace2"},
-		Spec: gwv1alpha2.ReferenceGrantSpec{
-			From: []gwv1alpha2.ReferenceGrantFrom{{
-				Group:     "gateway.networking.k8s.io",
-				Kind:      "HTTPRoute",
-				Namespace: "namespace1",
-			}},
-			To: []gwv1alpha2.ReferenceGrantTo{{
-				Group: "",
-				Kind:  "Service",
-				Name:  &backendName,
-			}},
-		},
-	}
-
-	client.EXPECT().
-		GetReferenceGrantsInNamespace(gomock.Any(), gomock.Any()).
-		Return([]gwv1alpha2.ReferenceGrant{refGrant}, nil)
-
-	resolver.EXPECT().
-		Resolve(gomock.Any(), route.GetNamespace(), gomock.Any()).
-		Return(&service.ResolvedReference{Type: service.ConsulServiceReference, Reference: &service.BackendReference{}}, nil)
-
-	require.NoError(t, route.Validate(context.Background()))
 }
 
 func TestRouteResolve(t *testing.T) {
