+++ internal/k8s/reconciler/utils.go
@@ -121,91 +121,6 @@ func routeAllowedForListenerNamespaces(ctx context.Context, gatewayNS string, al
 	return false, nil
 }
 
-// routeAllowedForBackendRef determines whether the route is allowed
-// for the backend either by being in the same namespace or by having
-// an applicable ReferenceGrant in the same namespace as the backend.
-//
-// TODO This func is currently called once for each backendRef on a route and results
-//   in fetching ReferenceGrants more than we technically have to in some cases
-func routeAllowedForBackendRef(ctx context.Context, route Route, backendRef gwv1alpha2.BackendRef, c gatewayclient.Client) (bool, error) {
-	fromNS := route.GetNamespace()
-	fromGK := metav1.GroupKind{
-		Group: route.GroupVersionKind().Group,
-		Kind:  route.GroupVersionKind().Kind,
-	}
-
-	toName := string(backendRef.Name)
-	toNS := ""
-	if backendRef.Namespace != nil {
-		toNS = string(*backendRef.Namespace)
-	}
-
-	// Kind should default to Service if not set
-	// https://github.com/kubernetes-sigs/gateway-api/blob/ef773194892636ea8ecbb2b294daf771d4dd5009/apis/v1alpha2/object_reference_types.go#L105
-	toGK := metav1.GroupKind{Kind: "Service"}
-	if backendRef.Group != nil {
-		toGK.Group = string(*backendRef.Group)
-	}
-	if backendRef.Kind != nil {
-		toGK.Kind = string(*backendRef.Kind)
-	}
-
-	return referenceAllowed(ctx, fromGK, fromNS, toGK, toNS, toName, c)
-}
-
-// referenceAllowed checks to see if a reference between resources is allowed.
-// In particular, references from one namespace to a resource in a different namespace
-// require an applicable ReferenceGrant be found in the namespace containing the resource
-// being referred to.
-//
-// For example, a Gateway in namespace "foo" may only reference a Secret in namespace "bar"
-// if a ReferenceGrant in namespace "bar" allows references from namespace "foo".
-func referenceAllowed(ctx context.Context, fromGK metav1.GroupKind, fromNamespace string, toGK metav1.GroupKind, toNamespace, toName string, c gatewayclient.Client) (bool, error) {
-	// Reference does not cross namespaces
-	if toNamespace == "" || toNamespace == fromNamespace {
-		return true, nil
-	}
-
-	// Fetch all ReferenceGrants in the referenced namespace
-	refGrants, err := c.GetReferenceGrantsInNamespace(ctx, toNamespace)
-	if err != nil || len(refGrants) == 0 {
-		return false, err
-	}
-
-	for _, refGrant := range refGrants {
-		// Check for a From that applies
-		fromMatch := false
-		for _, from := range refGrant.Spec.From {
-			if fromGK.Group == string(from.Group) && fromGK.Kind == string(from.Kind) && fromNamespace == string(from.Namespace) {
-				fromMatch = true
-				break
-			}
-		}
-
-		if !fromMatch {
-			continue
-		}
-
-		// Check for a To that applies
-		for _, to := range refGrant.Spec.To {
-			if toGK.Group == string(to.Group) && toGK.Kind == string(to.Kind) {
-				if to.Name == nil || *to.Name == "" {
-					// No name specified is treated as a wildcard within the namespace
-					return true, nil
-				}
-
-				if gwv1alpha2.ObjectName(toName) == *to.Name {
-					// The ReferenceGrant specifically targets this object
-					return true, nil
-				}
-			}
-		}
-	}
-
-	// No ReferenceGrant was found which allows this cross-namespace reference
-	return false, nil
-}
-
 func toNamespaceSet(name string, labels map[string]string) klabels.Labels {
 	// If namespace label is not set, implicitly insert it to support older Kubernetes versions
 	if labels[NamespaceNameLabel] == name {
